requirements.txt
nicegui
yt-dlp
faster-whisper
google-generativeai
pystray
Pillow
psutil
torch
nvidia-ml-py
